requirements.txt
openai
streamlit
google_genai
dotenv